Initial production release - BrightNest Academy

+++ shrishail-academy/src/main/java/com/shrishailacademy/repository/PaymentRepository.java
@@ -6,6 +6,7 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Optional;
 
@@ -31,10 +32,10 @@ public interface PaymentRepository extends JpaRepository<Payment, Long> {
     boolean existsByUserIdAndCourseIdAndStatus(Long userId, Long courseId, Payment.Status status);
 
     @Query("SELECT SUM(p.amount) FROM Payment p WHERE p.status = 'SUCCESS'")
-    Double getTotalRevenue();
+    BigDecimal getTotalRevenue();
 
     @Query("SELECT SUM(p.amount) FROM Payment p WHERE p.user.id = :userId AND p.status = 'SUCCESS'")
-    Double getTotalPaidByUser(@Param("userId") Long userId);
+    BigDecimal getTotalPaidByUser(@Param("userId") Long userId);
 
     @Query("SELECT p.paymentMethod, COUNT(p), SUM(p.amount) FROM Payment p WHERE p.status = 'SUCCESS' GROUP BY p.paymentMethod")
     List<Object[]> getPaymentMethodStats();
